backend/requirements.txt
numpy
sounddevice
openai-whisper
webrtcvad
torch
PyYAML
rapidfuzz
transformers
caldav
ics
jsonschema
blobfile
protobuf
SentencePiece